@@ -1,7 +1,7 @@
 {
   "name": "functions",
   "scripts": {
-    "lint": "eslint --ext .js,.ts .",
+    "lint": "eslint \"src/**/*.{ts,js}\" --max-warnings=0",
     "build": "tsc",
     "watch": "tsc --watch",
     "serve": "npm run build && firebase emulators:start --only functions",
@@ -16,24 +16,32 @@
   "main": "lib/index.js",
   "dependencies": {
     "@notionhq/client": "^2.2.15",
-    "axios": "^0.22.0",
+    "axios": "^1.7.9",
     "date-fns": "^2.25.0",
-    "firebase-admin": "^10.0.2",
-    "firebase-functions": "^3.16.0",
+    "dotenv": "^17.2.1",
+    "firebase-admin": "^13.1.0",
+    "firebase-functions": "^6.4.0",
+    "lodash": "^4.17.21",
     "request": "^2.88.2",
     "slugify": "^1.6.5",
     "xml-js": "^1.6.11"
   },
   "devDependencies": {
+    "@eslint/compat": "^1.3.2",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.35.0",
+    "@types/lodash": "^4.17.13",
     "@types/request": "^2.48.5",
     "@types/ws": "8.5.4",
-    "@typescript-eslint/eslint-plugin": "^3.9.1",
-    "@typescript-eslint/parser": "^3.8.0",
-    "eslint": "^7.6.0",
+    "@typescript-eslint/eslint-plugin": "^8.43.0",
+    "@typescript-eslint/parser": "^8.43.0",
+    "eslint": "^9.35.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-import": "^2.22.0",
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-import": "^2.31.0",
     "firebase-functions-test": "^0.3.3",
-    "typescript": "^4.1.6"
+    "globals": "^16.4.0",
+    "typescript": "^5.8.3"
   },
   "private": true
 }
