@@ -1,8 +1,8 @@
-import * as functions from 'firebase-functions';
 import axios from 'axios';
+import { differenceInMinutes, isWithinInterval, parse, parseISO } from 'date-fns';
+import { HttpsError, onCall } from 'firebase-functions/v2/https';
 import { xml2json } from 'xml-js';
 import { Channel, Programme, Tv } from './model/tv.model';
-import { differenceInMinutes, isWithinInterval, parse, parseISO } from 'date-fns';
 
 const PROGRAM_URL = 'https://xmltv.ch/xmltv/xmltv-tnt.xml';
 
@@ -17,7 +17,7 @@ type MappedProgram = Omit<Programme, '_attributes'> & {
 
 const getAllPrograms = async (): Promise<MappedProgram[]> => {
 	const response = await axios.get<string>(PROGRAM_URL);
-	const json: { tv: Tv } = JSON.parse(xml2json(response.data, { compact: true }));
+	const json = JSON.parse(xml2json(response.data, { compact: true })) as { tv: Tv };
 	const channels = json.tv.channel;
 	const programs = json.tv.programme;
 
@@ -43,23 +43,23 @@ const getAllPrograms = async (): Promise<MappedProgram[]> => {
 	});
 };
 
-export const primeTimePrograms = functions
-	.region('europe-west1')
-	.https.onCall(async (data, context): Promise<MappedProgram[]> => {
-		const uid = context.auth?.uid;
+export const primeTimePrograms = onCall<{ targetRange: { from: Date; to: Date } }>(
+	async (request): Promise<MappedProgram[]> => {
+		const uid = request.auth?.uid;
 		if (!uid) {
-			throw new functions.https.HttpsError('failed-precondition', 'not_authenticated');
+			throw new HttpsError('failed-precondition', 'not_authenticated');
 		}
 
 		const programs = await getAllPrograms();
 
 		return programs.filter((program) => {
 			const programStart = parseISO(program._attributes.start);
-			const targetRangeFrom = parseISO(data.targetRange?.from);
-			const targetRangeTo = parseISO(data.targetRange?.to);
+			const targetRangeFrom = parseISO(request.data.targetRange?.from.toString());
+			const targetRangeTo = parseISO(request.data.targetRange?.to.toString());
 			return (
 				program._attributes.durationHours > 0.75 &&
 				isWithinInterval(programStart, { start: targetRangeFrom, end: targetRangeTo })
 			);
 		});
-	});
+	}
+);
