@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
-import { tap } from 'rxjs/operators';
-import { AppStore } from './app.store';
-import { UserData } from '../app/model/user-data';
-import { UserConfig } from '../app/model/user-config';
-import { addWeeks, isBefore } from 'date-fns/esm';
 import { doc, docData, Firestore, updateDoc } from '@angular/fire/firestore';
-import { DocumentReference } from 'rxfire/firestore/interfaces';
-import { UntilDestroy } from '@ngneat/until-destroy';
 import { UpdateData } from '@firebase/firestore';
+import { UntilDestroy } from '@ngneat/until-destroy';
+import { addWeeks, isBefore } from 'date-fns/esm';
+import { tap } from 'rxjs/operators';
+import { UserConfig } from '../app/model/user-config';
+import { UserData } from '../app/model/user-data';
+import { getFirestoreConverter } from '../utils/firestore-converter';
+import { AppStore } from './app.store';
 
 export type ShiftDirection = 'previous' | 'next';
 
@@ -20,14 +20,16 @@ export class AppService {
 	) {}
 
 	fetchConfig(uid: string) {
-		return docData(doc(this.firestore, `/users/${uid}`) as DocumentReference<UserData>).pipe(
-			tap((userData) => this.appStore.update({ userData }))
-		);
+		return docData(
+			doc(this.firestore, `/users/${uid}`).withConverter(getFirestoreConverter<UserData>())
+		).pipe(tap((userData) => this.appStore.update({ userData })));
 	}
 
 	setConfig(config: UpdateData<UserConfig>) {
 		const uid = this.appStore.getValue().user!.uid;
-		const userDoc = doc(this.firestore, `/users/${uid}`) as DocumentReference<UserData>;
+		const userDoc = doc(this.firestore, `/users/${uid}`).withConverter(
+			getFirestoreConverter<UserData>()
+		);
 		return updateDoc(userDoc, { config });
 	}
 
